- Fixed Namespacing for ‘.menu’ to prevent ionic conflict
- Toggles now work, removed some css from the main stylesheet that was
from an old demo that was conflicting with Ionc and causing issues

- Updated some core app settings to prevent “App Bounce” when you come
to the top or bottom of scrolling. This can get annoying in things like
the flyout menu. NOTE: Scrolling in Safari Desktop seems to have been
disabled now, but works great in XCODE and Sim

- Looking into Simplepie Sorting (They say its sort of a bug, but there
is a hack.) checking it out

- Looking into Load More with RSS. Have to see if simple pie has the
ability to “get the next set of articles”. Issue here may be that the
next set may not be “in order” (due to the previous bug and other
things)… Maybe for now we can just have it pull more at the get go…
more soon